LayoutTests/fast/frames/seamless/seamless-contenteditable-not-inherited.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../js/resources/js-test-pre.js"></script>
    <script>
        window.onload = function () {
            debug("This test ensures that content inside a seamless iframe does not inherit editability via the contenteditable attribute on a parent element, but does via a CSS rule that cascades into the frame.");

            window.span = document.querySelector('div > iframe').contentDocument.querySelector('span');
            window.p = document.querySelector('body > iframe').contentDocument.querySelector('p');

            shouldBeEqualToString("window.getComputedStyle(span).getPropertyCSSValue('-webkit-user-modify').cssText", "read-only");
            shouldBeEqualToString("window.getComputedStyle(p).getPropertyCSSValue('-webkit-user-modify').cssText", "read-write");
        };
    </script>
</head>
<body>
    <div contenteditable>
        <iframe seamless srcdoc="<span>This span is not editable.</span>"></iframe>
    </div>
    <style>
        p { -webkit-user-modify: read-write; }
    </style>
    <iframe seamless srcdoc="<p>This paragraph is not editable.</p>"></iframe>
</body>
</html>
